+++ projects/ui-framework/bob-charts/src/charts/combinations/bar-line-chart/bar-line-chart.component.ts
@@ -9,15 +9,7 @@ import {
 import { ChartCore } from '../../chart/chart-core';
 import { SeriesOptionsType } from 'highcharts';
 import { ChartTypesEnum } from '../../charts.enum';
-
-export const COLUMN_BAR_WIDTH = {
-  pointPadding: 0.1,
-  groupPadding: 0.08,
-};
-export const MULTI_COLUMN_WIDTH = {
-  pointPadding: 0.1,
-  groupPadding: 0.2,
-};
+import { COLUMN_BAR_WIDTH, MULTI_COLUMN_WIDTH } from '../../charts.const';
 
 @Component({
   selector: 'b-bar-line-chart',
